#recovered-old-smart-contracts
#workflow
I agreed @Marcelo's suggestion, I am going to upgrade smart contracts from old code based on his feedback.

# @Marcelo's suggestion:
1) workflow
I know that you have been developing solo up until now (with Omar's oversight) but in my experience it's a good practice to keep 1 main branch (protected if possible) and then iterate on a secondary one.
this is a big help in more than 1 way: it prevents potentially unwanted updates to the main branch, but more importantly it allows us to do PRs. PRs are a great interface to submit reviews on and give an opportunity to apply further changes to the code before it is merged to the main branch. it also makes updates to the repo more readable and traceable.
what do you think of taking this approach for next time?