#fixed-events-for-oldxxx-field
In @Marecelo's feedback,
i'll provide some written feedback but we can also set up a call if it'd be helpful
4:58
so, right now the contract is past the allowed limit in KiB, which is a very common problem. I think in the long run it will be have to be split into multiple contracts, because having the size so close to the limit will cause the problem to occur again after applying even small changes to the code.
having said that, here are some ideas that can help you go below the limit as it currently stands:
5:01
1. there are a lot of update events which emit both the old value as well as the new, updated value. this isn't strictly necessary as long as we emit only the new value with each event. if you want to put together a historical list of the values as they change, you can simply filter the events by the fixed parameter and then read each new value. as it is right now, we have redundant information, like this:
// value is at A, and then it is changed to B
emit valueUpdated(A, B)
// value is at B, and then it is changed to C
emit valueUpdated(B, C)
here you can see that "B" is repeated, if we only had this instead:
emit valueUpdated(A) // maybe this is the initial set and not an update, doesn't matter ultimately
//...
emit valueUpdated(B)
emit valueUpdated(C)
you can still query the chain and reconstruct the history A -> B -> C, but the events are smaller in size because they have 1 less field