#remove-TIExShareCollectionExists
In @Marcelo's feedback:

- do we really need the shareCollectionExists mapping? couldn't we check for a zero-value on the shareCollections mapping? it seems redundant at a first glance

- the important thing here is that every type in solidity has a zero-value. for uint256, it is the number 0, for address it is address(0) and so-on.
in the code that you've provided, both values will start at 0 by default. what you need to ask yourself here is: does this make sense from a business logic point of view?
let's take the following example:
uint256 balance;
will this value start at 0? yes
can this value be 0 and still be initialized? also yes, since conceptually you can have 0 balance and this doesn't mean that a struct that has this variable wasn't initialized.
there's no need to check for all values to be their respective zero-value in order to be sure that the we are in the case of "non-initialized" or "non-existent", but it really depends on the context.
let's take this to the TIExShareCollection struct:
    struct TIExShareCollection {
        /// @notice Indicates the maximum share supply.
        uint256 maxSupply;
        /// @notice Indicates the total investment in the share
        uint256 totalInvestment;
        /// @notice Indicates the amount that has been withdrawn.
        uint256 withdrawnAmount;
        /// @notice Indicates the price per share.
        uint256 price;
        /// @notice Indicates the time when the share collection was launched.
        uint256 launchStartTime;
        /// @notice Indicates the maximum share purchase allowed per account.
        uint256 maxSharePurchase;
        /// @notice Indicates whether the sale is paused or not.
        bool paused;
        /// @notice Indicates whether the share collection is available for
        /// trading and share sale or not.
        bool blocked;
        /// @notice Indicates whether the share collection is only available
        /// for U.S. investors or not.
        bool forOnlyUSInvestors;
    }
out of all of these values, I can think of at least 1 that can never be zero if the structure has valid data, and that is the timestamp (launchStartTime )
when you create this struct, that field is set to block.timestamp. therefore, if launchStartTime == 0, then we are in the case of a "non-existing share collection". you don't need to check any other value, this is your logical condition for "structure exists".