#remove-repeated-mappings
In @Marcelo's feedback:

repeated mappings. there are 3 or 4 mappings that go from "model ID" to something else, either an index, or the metadata. you can merge all of these mapping into a single one, that goes from model ID to a struct. this struct can nest the uint256 fields as well as the metadata structure. you'll have to be careful with this change since it will impact lots of places in the code, but the tests should let you know if something is wrong.

    /// MAIN

    /// @notice Mapping model id to TIEx Model
    mapping(uint256 => TIExModel) public TIExModels;

    /// OTHERS

    /// @notice Mapping from model ID to index of the creator models list
    mapping(uint256 => uint256) private _ownedModelsIndex;

    /// @notice Mapping from model id to position in the allModels array
    mapping(uint256 => uint256) private _allModelsIndex;

    /// @notice Mapping model id to contriuted model list
    mapping(uint256 => Contribution[]) public contributedModels;