#contract-size-issue
maybe, since the mint function is already paired with the buyShares function, it would be easier to extract the model data + metadata to a separate contract and keep the 1155 with the shares code.
my only concern is that there's enough code there to extract.
why don't we try listing all of the functions, fields, structs, events and inheritance for each contract in this alternative?